Init MapObject Struct
* Converted MapObject class and interface into a struct for equating different MapObjects.
